splitted wsdiscovery into multiple files
